Education is when you read the fine print. Experience is what you get if you don't
