+++ 0x11-C_printf/0-print_character_string.c
@@ -1,14 +1,5 @@
 #include "main.h"
-#include <stdarg.h>
-#include <unistd.h>
-/**
- * handle_binary - Converts an unsigned integer to binary and prints it
- * @args: A va_list of arguments passed to the calling function
- *
- * Return: The total number of characters printed,
- * including the converted number
- */
-int handle_binary(va_list args);
+
 /**
  * _printf - produces output according to a format
  * @format: character string containing format specifier
@@ -65,32 +56,20 @@ int _printf(const char *format, ...)
 						printed_chars = handle_binary(args);
 						break;
 					}
+				case 'd':
+				case 'i':
+					{
+						handle_integer(args, printed_chars);
+						break;
+					}
+				case '\0':
+					{
+						va_end(args);
+						return (-1);
+					}
 			}
 		}
 	}
 	va_end(args);
 	return (printed_chars);
-}
-/**
- * handle_binary - Converts an unsigned integer to binary and prints it
- * @args: A va_list of arguments passed to the calling function
- * Return: The total number of characters printed,
- * including the converted number
- */
-int handle_binary(va_list args)
-{
-	char buffer[33];
-	unsigned int num;
-	int i;
-
-	num = va_arg(args, unsigned int);
-	i = 0;
-	while (num)
-	{
-		buffer[i++] = (num % 2) + '0';
-		num >>= 1;
-	}
-	while (i--)
-		write(1, &buffer[i], 1);
-	return (i);
 }
